refactor: split expression pf planner into one part.
